fix: some small bugs
1) issue on flz datasize computation
2) size of message and empty plc get var
3) restarting the spooler should wait or we get the message "No spooler connected"
4) cluster mock should manage rotation of blocks

+++ rx_common_lib/rx_flz.c
@@ -360,7 +360,7 @@ int flz_load(SPageId *id, const char *filedir, const char *filename, int printMo
 				}
 			}
 			pinfo->lengthPx += 2*wul;
-			pinfo->dataSize = pinfo->lengthPx * pinfo->lineLen; 
+            pinfo->dataSize = (UINT64)pinfo->lengthPx * pinfo->lineLen; 
 
             //  TrPrintfL(TRUE, "DECOMPRESSING >>%s<<, page %d, time=%d ms", filepath, id->page, rx_get_ticks()-time);
 		}
